Move mocking of Grakn constructs to the appropriate location
