+++ kglib/utils/grakn/test/mock/answer.py
