+++ kglib/utils/grakn/test/mock/concept.py
