+++ kglib/utils/graph/create/queries_to_graph_it.py
@@ -22,8 +22,8 @@ import networkx as nx
 from grakn.client import GraknClient
 
 from kglib.utils.graph.create.queries_to_graph import build_graph_from_queries
-from kglib.utils.graph.test.mock.answer import MockConceptMap
-from kglib.utils.graph.test.mock.concept import MockType, MockAttributeType, MockThing, MockAttribute
+from kglib.utils.grakn.test.mock.answer import MockConceptMap
+from kglib.utils.grakn.test.mock.concept import MockType, MockAttributeType, MockThing, MockAttribute
 from kglib.utils.graph.test.case import GraphTestCase
 from kglib.kgcn.core.ingest.traverse.data.context.neighbour import Thing, build_thing
 from kglib.kgcn.test.base import GraknServer
